New test CopyPagesWithExtras with test method testAnnotationSampleStandard and test data AnnotationSample.Standard.pdf.
This test copies the single page from the output of the PdfClown
AnnotationSample into a new PDF.

As it turns out, the page copying mechanism currently fails to see
inherited properties. Thus, this copy operation takes place without
MediaBox and without Resources.

All the annotations, on the other hand, for which this test was written
in the first place, seem to have been copied correctly.