New test ChangeSignatureAppearance focusing on the stackoverflow question
How to associate a previous signature in a new signature field
http://stackoverflow.com/questions/37027579/how-to-associate-a-previous-signature-in-a-new-signature-field
BLANK-signed.pdf
a blank file from elsewhere with an invisible signature

Quite surprisingly it turns out that changing the signature appearance
is possible without breaking the signature, merely a warning appears
which can be hidden by simply signing again.

This is the equivalent of the iText5 test area test with the same name.