New ChangeSignatureAppearance test testChangeAppearancesWithNameAllPages focusing on the stackoverflow question
How to associate a previous signature in a new signature field
http://stackoverflow.com/questions/37027579/how-to-associate-a-previous-signature-in-a-new-signature-field
monitor_2016_99.pdf
http://185.49.12.119/~pogdan/7spacedot/monitor_2016_99.pdf
which is a sample file from another question.
http://stackoverflow.com/questions/37439613/itextpdf-insert-space-beetwen-7-and-dot-after-extract-text

Similarly to testChangeAppearancesWithName(), this test adds signature
appearances not breaking signature validity containing the signer
certificate subject common name, but this time it adds an appearance to
each page.

The sample file already has an appearance on one page. This appearance
is forcefully removed.