New test TwoStepSigning focusing on the stackoverflow question
In Itext 7, how to sign a pdf with 2 steps?
http://stackoverflow.com/questions/39151230/in-itext-7-how-to-sign-a-pdf-with-2-steps
document_in.pdf
https://1drv.ms/b/s!AkF6t4TavwMvdLmiyQlYVcCLqw4

Indeed, the result of this PreSign code is broken for the sample file.
The reason is that in parallel to the original file iText tries to use
object streams as much as possible, unfortunately even for the signature
dictionary which is devastatingly wrong.