New test AddPageAndSign focusing on the stackoverflow question
Adding a new page to PDF and create signature with iText 7
http://stackoverflow.com/questions/39296225/adding-a-new-page-to-pdf-and-create-signature-with-itext-7

The original code used "SunJSSE" as provider argument to the sign call.
Using that setting the signing procedure here fails with a
java.security.NoSuchAlgorithmException: no such algorithm: SHA1 for
provider SunJSSE. Changing the provider argument to "BC" creates a
properly certified PDF.

Added test key store in keystores/ for this and similar signing tests.