New MoveSectionCleanly test testMoveCleanSectionBody focusing on the stackoverflow question
Moving text on a page with iText7 retaining font, color, style, … but changing size of the text
http://stackoverflow.com/questions/40551553/moving-text-on-a-page-with-itext7-retaining-font-color-style-but-changing

This example shows how to "move" content from a source rectangle to a target rectangle without the text being still selectable in the source rectangle.

The method moveCleanSection is the actual worker method here.