New class PdfCanvasEditor focusing on the stackoverflow question
Manipulate paths, color etc. in iText
http://stackoverflow.com/questions/40951776/manipulate-paths-color-etc-in-itext

This class presents a simple content stream editing framework. As is it creates an equivalent copy of the original page content stream. To actually edit, simply overwrite the method write(PdfCanvasProcessor, PdfLiteral, List) to not (as in this class) write the given operations as they are but change them in some fancy way.

This is a port of the iText 5 test area class PdfContentStreamEditor.

Additionally there are new EditPageContent tests which are ports of tests for the original iText 5 class:

* testRemoveBoldMTTextDocument shows how to remove text matching the filter condition given by the OP, i.e. any text drawn using a font whose name ends with "BoldMT".
* testRemoveBigTextDocument shows how to remove text filtered by font size.
* testChangeBlackTextToGreenDocument shows how to change the color of text of a given color. In this case, black text is changed to green.

These tests operate on the PDF

document.pdf
https://1drv.ms/b/s!AmNST-TRoPSemi2k0UnGFsjQM1Yt

which had been provided by the OP of 

How to remove filtered content from a PDF with iText
http://stackoverflow.com/questions/38498431/how-to-remove-filtered-content-from-a-pdf-with-itext