New EnableLtv test focusing on the stack overflow question
iText 7 adding ltv to existing signature
http://stackoverflow.com/questions/42899185/itext-7-adding-ltv-to-existing-signature

This test applies the method addLtvNoTS() to a simple D-Trust signed PDF with no embedded LTV information. That method does only add LTV information for the immediate PDF signatures and document timestamps, not for already embedded or now added OCSP response or CRL signatures. As D-Trust OCSP responses are signed by certificates which need own LTV information while the CRL is signed by a signature which doesn't, we have to use CRLs all over here to get a "LTV-enabled" document.

Furthermore we explicitly have to give a link to the CRL as D-Trust mentions two CRL distribution points, the first one being reachable via LDAP, and iText only uses the first distribution point but does not support LDAP.

addLtvNoTS() is the iText 7 equivalent to the iText 5 `addLtvNoTS` the OP refers to in a comment and which originally was posted on stack overflow as the answer <http://stackoverflow.com/a/27906587/1729265>.