New ExtractTextPlusY test focusing on the stack overflow question
How to get the text position from the pdf page in iText 7
http://stackoverflow.com/questions/43746884/how-to-get-the-text-position-from-the-pdf-page-in-itext-7

testExtractTextPlusYFromTest shows how to extract text with its characters' respective y coordinates from a document. It makes use of TextPlusYExtractionStrategy, a strategy derived from the LocationTextExtractionStrategy tweaked to also return a TextPlusY instance which contains text plus y coordinates.

Beware, this is but a proof-of-concept which in particular assumes text to be written horizontally, i.e. using an effective transformation matrix with b and c equal to 0.
Furthermore the character and coordinate retrieval methods of TextPlusY are not at all optimized and might take long to execute.