pdfHTML: Responsive Design from iText web site
As the links on the iText web site for the pdfHTML: Responsive Design example <http://developers.itextpdf.com/content/itext-7-examples/converting-html-pdf/pdfhtml-responsive-design> don't work anymore and the files could not be found on <https://git.itextsupport.com/>, here an adapted version.

It can be run with iText 7.0.2 and html2pdf 1.0.0.