New CheckForOverlappingText tests focusing on the stack overflow question
How to find whether a PDF file has overlapping text or not, using c#
https://stackoverflow.com/questions/46362605/how-to-find-whether-a-pdf-file-has-overlapping-text-or-not-using-c-sharp

New event listener OverlappingTextSearchingStrategy which attempts to check whether overlapping text as defined by the OP exists on the page parsed through it.
The OP defined "overlapping text" as text whose bounding box overlaps the bounding box of other text or of some image. Neither vector graphics nor annotations are to be considered.
This is a prove-of-concept only. In particular it works best by far for horizontal and vertical text and images as the bounding boxes are used  with only horizontal and vertical edges.
Kerning and tightly set text might result in false positives. To make this less likely, text bounding boxes are slightly shrunk before they are stored for later comparison. The margins used may have to be fine-tuned for the task at hand.

testSampleCMYK() tests the OverlappingTextSearchingStrategy against a document with very tightly set text. As expected there are many false positives.

testInterinosDisponibles0590_20170915() tests the OverlappingTextSearchingStrategy against a document with loosely set text. As expected there are no false positives.

testOverlapTestPdf() tests the OverlappingTextSearchingStrategy against a document with overlapped text. As expected the strategy recognizes overlapped text.

testNonOverlapTestPdf() tests the OverlappingTextSearchingStrategy against a document without overlapped text. As expected the strategy recognizes no overlapped text.