New BreakByHandling test focusing on the stack overflow question
iText 7.0.4.0 - PdfWriter produces corrupted PDF for certain PDF file inputs
https://stackoverflow.com/questions/46479449/itext-7-0-4-0-pdfwriter-produces-corrupted-pdf-for-certain-pdf-file-inputs
sample1_input.pdf
https://drive.google.com/open?id=0B3NPOZswWocQV09KMW5fbFVyUm8

The issue can be reproduced in iText 7.0.4 but not anymore in 7.0.5 SNAPSHOT, and there indeed is a commit described as "Fix bugs in pages tree rebuilding" in a block described as "handle mix of PdfPage and PdfPages". The PDF in question does have a page tree node with mixed child node types, Page and Pages, and the latter subtree even does not contain any Page descendants.