New AddSignatureField test focusing on the stack overflow question
Adding a signature form field to a page it's visible in the last page also
https://stackoverflow.com/questions/46906725/adding-a-signature-form-field-to-a-page-its-visible-in-the-last-page-also

Indeed, the signature appears on the first and on the last page.

This is caused by the use of PdfAcroForm.addField(PdfFormField) which is documented to add the field on the last page. The OP should use PdfAcroForm.addField(PdfFormField, PdfPage) instead.