New StampNoChange tests focusing on the stack overflow question
Append mode requires a document without errors, even if recovery is possible
https://stackoverflow.com/questions/47281240/append-mode-requires-a-document-without-errors-even-if-recovery-is-possible
word.pdf
https://www.dropbox.com/s/i7eeamw9xouf76l/word.pdf?dl=0

testAppendWord reproduces the issue observed by the OP: A first stamping seems to work but the second one throws an error. The cause is that the first stamping already created a broken result which makes the second one stumble.

testAppendWordSigning shows that the issue is not bound to the signing use case during which the OP observed it: Signing simply produces the same fault in the result document as arbitrary stamping does in the test testAppendWord.

testNoAppendWord shows that the issue is related to the append mode: Without appending everything works just fine.