New CreateFormInTable tests focusing on the stack overflow question
iText 7.0.5 - Forms - NullPointerException when adding a lot of tables with form fields
https://stackoverflow.com/questions/47594075/itext-7-0-5-forms-nullpointerexception-when-adding-a-lot-of-tables-with-form

Indeed, the issue can be reproduced as testCreateFormInTable shows.

testCreateFormInTableVariant actually reproduces the issue in a much simpler way.

The cause of the issue is that for a form field (widget) without an associated page, iText tries to determine the page that widget is on by iterating all document pages and looking up their respective annotation arrays.

In the code above, though, the previous newly created pages, at least the first one, has already been flushed to the writer and de-initialized in memory. Thus, the loop over the pages tries to retrieve a value from a de-initialized page dictionary, i.e. from a null Map, hence the NullPointerException.