New CreateCyrillicText test focusing on the stack overflow question
While parsing pdf with iText7 chars move on fixed interval (with Freeset font)
https://stackoverflow.com/questions/47677271/while-parsing-pdf-with-itext7-chars-move-on-fixed-interval-with-freeset-font
freeset.ttf
http://allfont.ru/download/freeset/

This code reproduces the issue. Apparently iText 7 does understand the font well enough to select the needed subset and reference the correct glyphs from the content but it has problems building an appropriate ToUnicode map. This is not a general problem, though, as the parallel test with Arial shows.