New CreateFormInTable tests focusing on the stack overflow question
iText 7.0.5 - Forms - NullPointerException when adding a lot of tables with form fields
https://stackoverflow.com/questions/47594075/itext-7-0-5-forms-nullpointerexception-when-adding-a-lot-of-tables-with-form

testCreateFormInTableNoFlush: Switching off immediate flushing prevents the exception but the result is not as expected by the OP: the table is spread over 4 pages but all form fields are added to the last page.

testCreateFormInTableVariantNoFlush: The simpler variant to reproduce the issue executed in the test method testCreateFormInTableVariant can be prevented from throwing an exception by not using immediate flush. In this case the result even looks as expected (in contrast to the result of the no-immediate-flush version of the original code, testCreateFormInTableNoFlush).