New CreateSpecialSignatureAppearance test focusing on the stack overflow question
Signature appearance font color in layer2 (itext 7)
https://stackoverflow.com/questions/48467637/signature-appearance-font-color-in-layer2-itext-7

Indeed, the option to transport color via the font in iText 5 seems not to have been taken into account when porting the PdfSignatureAppearance class to iText 7.

Of course there always is the option of creating any layer manually. But if one does not want to do so, one can tweak the layer content after making iText create the appearances; the latter requires reflection, though.