New RenderDivOverflowHidden test focusing on the stack overflow question
iText 7: How can I allow overflow in a Div?
https://stackoverflow.com/questions/49601728/itext-7-how-can-i-allow-overflow-in-a-div

This method tests the OverflowHiddenDivRenderer which renderer renders a Div so that it appears cut off at its HEIGHT property; in particular, if the cut off occurs in the middle of a line, that line is still visible above that cut off line.

Beware, I may well have forgotten some special cases. I think in particular of problems in context with rotated content (which in super.getOccupiedAreaBBox() is of influence) or area breaks (I don't set a next renderer in OverflowHiddenDivRenderer with an adapted height).