New ExtractPositionalText test focusing on the stack overflow question
How to find text position and boundary in iText 7
https://stackoverflow.com/questions/50894662/how-to-find-text-position-and-boundary-in-itext-7
chinese.pdf
https://github.com/voidmagic/pdf-parser/blob/master/data/chinese.pdf

As it turns out the cause is a bug in iText's CMap loading. When loading a font encoding CMap, iText stores the data inversely in a Map, i.e. akin to `map.put(value, key)`. Unfortunately the encoding of the footer font at hand is GBK-EUC-H which is not injective. Thus, during CMap loading some entries are overridden and the matching character codes cannot be resolved, causing extracted text containing replacement characters unnecessarily and extracted locations ignore these characters.