New ExtractAnnotationImage test focusing on the stack overflow question
How to retrieve the image of a PdfStampAnnotation
https://stackoverflow.com/questions/51087035/how-to-retrieve-the-image-of-a-pdfstampannotation
add_stamp.pdf
output of the iText test com.itextpdf.samples.sandbox.annotations.AddStamp

This test shows how to extract images used in annotation appearances.
Beware: This is a very simple extractor, it does not check whether the image really is visible in the annotation, and it ignores any masks, i.e. transparency.