New SignTwice tests focusing on the stack overflow questions
Multiple signature not working
https://stackoverflow.com/questions/51812973/multiple-signature-not-working
signed.pdf
https://drive.google.com/open?id=1dOZT_VGnqVe1II_DIiISGcD81be5V6i6

testSignSignedAgain shows that the OP's code indeed results in warnings that annotations have been modified.

testStampSigned reduces the issue to the inner cause: PdfPage.getAnnotations() as a side effect adds P entries to the annotation dictionaries, references to the page they're on, and marks them as changed.