New AddStampAnnotation test focusing on the stack overflow questions
itext7 stamp and redact annotation not working
https://stackoverflow.com/questions/51887277/itext7-stamp-and-redact-annotation-not-working
calendar_2018-08-04_2018-08-19.pdf
https://www.dropbox.com/s/3jh0e4ttr3q1ne9/calendar_2018-08-04_2018-08-19.pdf?dl=0

Indeed, the redaction and stamp annotations created by the OP are essentially transparent in the viewer and non-existing in a print-out.

This can be fixed as shown in the annotations link2x and stamp1x added to the OP's original annotations: By adding a normal appearance and by setting the PRINT flag.