New EditPageContent test focusing on the stack overflow question
itext7 - How to filter render events during write
https://stackoverflow.com/questions/51691625/itext7-how-to-filter-render-events-during-write
calendar_2018-08-04_2018-08-19.pdf
https://www.dropbox.com/s/3jh0e4ttr3q1ne9/calendar_2018-08-04_2018-08-19.pdf?dl=0

This test shows how to remove the lines "Created by:" and "Calendar:" from a PDF created from Google Calendar in the Agenda format.

For this the PdfCanvasEditor has been extended to allow peeking at the operation to come before it is processed by the original IContentOperator.

Beware, this is a proof-of-concept, not a final and complete solution.

In particular it is customized for the OP's use case: This PdfCanvasEditor only inspects and edits the first level form XObjects of each page because PDFs created from Google Calendar in the Agenda format contain all their page content in a form XObject which in turn is drawn in the page content stream. Furthermore, writing is expected to occur parallel to the top of the page.

Improving the class for general use remains as an exercise for the reader.