New AddWmfImage test focusing on the stack overflow question
Adding and rendering this WMF as a vector graphic in a new PDF using iText7
https://stackoverflow.com/questions/51844584/adding-and-rendering-this-wmf-as-a-vector-graphic-in-a-new-pdf-using-itext7
Image.wmf
http://s000.tinyupload.com/index.php?file_id=12631255103057077641

Rewriting the OP's code for streaming resource access like here actually showed up another issue: The WmfImageData constructor accepting a byte[] throws a NPE because it tries to determine the file type using its URL which doesn't exist.

After fixing the issue above and running the code, one can indeed witness an issue in the WMF import or in the given WMF file.