New RegularExpressionTest test focusing on the stack overflow question
iText PDFSweep RegexBasedCleanupStrategy not work in some case
https://stackoverflow.com/questions/52635416/itext-pdfsweep-regexbasedcleanupstrategy-not-work-in-some-case

Indeed, word boundaries at the end of lines might not be visible, cf. the search for "\\beinigen\\b" here.

This can be fixed in the CharacterRenderInfo method mapString: In the else-block of its `if (chunk.sameLine(lastChunk))` structure first execute a sb.append(' ') call.

As an aside updated pdfSweep version to 2.0.3-SNAPSHOT and pdfHtml version to 2.1.1-SNAPSHOT and applied follow-up fixes.