New AddPagenumberToCopy test focusing on the stack overflow question
Adding page number text to pdf copy gets flipped/mirrored with itext 7
https://stackoverflow.com/questions/52743566/adding-page-number-text-to-pdf-copy-gets-flipped-mirrored-with-itext-7
li.persia.pdf
https://drive.google.com/open?id=11_9ptuoRqS91hI3fDcs2FRsIUEiX0a84

Using the OP's original code in testLikeAibanez the generated page numbers indeed appear upside down somewhere in the middle of the page.

In testLikeAibanezFixed the generated page numbers appear in the upper right corner of the pages without being fliped or rotated.
The fix is to use a PdfCanvas constructed with the wrapOldContent parameter set to true.