New AddImageUnder tests focusing on the stack overflow question
How to bring image to the front(of the text/image) or send the image to
the back((of the text/image)) in IText7 using java?
https://stackoverflow.com/questions/53231903/how-to-bring-image-to-the-frontof-the-text-image-or-send-the-image-to-the-back

testAddLikeKishorePenmetsa: Indeed, Adobe Reader complains about an error on the page. The cause is the use of a throw-away resources object. It causes the image resource to not end up in the PDF at all.
testAddLikeKishorePenmetsaFixed: Using the page resources instead of a throw-away resources object fixes the issue. Furthermore, the dangling saveState instruction has been removed.