New CheckForCharacter test focusing on the stack overflow question
How to find out what letter is not in the font?
https://stackoverflow.com/questions/53546009/how-to-find-out-what-letter-is-not-in-the-font
ArtScript.ttf as "ArtScript-a-emptied.ttf" and with 'a' actually removed
as "ArtScript-a-removed.ttf"
https://dropmefiles.com/8lFiM

This test checks how PdfFont deals with missing characters in a font.

The OPs original font program "ArtScript-a-emptied.ttf" merely has
cleaned the glyph for 'a' void of any drawing instructions, the
manipulated font program "ArtScript-a-removed.ttf" has actually removed
the glyph for 'a'.

PdfFont.containsGlyph(int) reports accordingly.