New SetButtonImage test focusing on the stack overflow question
iText 7 : Set image to button that appears multiple times
https://stackoverflow.com/questions/54107835/itext-7-set-image-to-button-that-appears-multiple-times
itext_multiple_images.pdf
https://drive.google.com/file/d/1k1hctPecvwMQ-2eX1ieoRDEP5ktBrYWb/view?usp=sharing

Indeed, setting the image of a multi-widget pushbutton field fails.
Furthermore, version 7.1.4 used to throw an exception during form
flattening which already is fixed in the 7.1.5-SNAPSHOT as of now.

The cause of the error is the lacking support for form fields with
multiple widgets of PdfFormField.regenerateField().