New DetermineFormFieldFonts test focusing on the stack overflow question
Getting form field font information in itext7
https://stackoverflow.com/questions/54455849/getting-form-field-font-information-in-itext7
TextFieldsWithFonts.pdf, a PDF with several text fields with different
fonts

This test shows how to determine form field font information. This
actually uses an iText method for doing just that but as the method is
protected, a small trick must be used.