New InsertInSpace test focusing on the stack overflow question
IText 7 + Add source PDF content to destination PDF
https://stackoverflow.com/questions/56522616/itext-7-add-source-pdf-content-to-destination-pdf

testInsertSimpleTestPdf reduces the page size of the source document page to its bounding box which is determined using the MarginFinder content parsing event listener. Then it forwards that page to insertion into the OP's document with header and footer generation.

MarginFinder is a port of the iText5 test area render listener MarginFinder. It is here used to determine the area of the OP's source PDF page with content.