New CorrectSizes test focusing on the stack overflow question
Why are more Pixel used in resulting PDF than in source html?
https://stackoverflow.com/questions/57476868/why-are-more-pixel-used-in-resulting-pdf-than-in-source-html

This essentially is the OP's code. Indeed, the difference in image width can be reproduced: iText renders at 72dpi, browsers at 96dpi. As the web page does not have a "background-size" set, this difference actually is ok: In this case the intrinsic size is used but the spec does not tell how to determine this "intrinsic size".

For this test org.jsoup:jsoup had to be added as new dependency.
Also switched from 7.1.7-SNAPSHOT to 7.1.8-SNAPSHOT.