New ChangeColor test focusing on the stack overflow question
Traverse whole PDF and change blue color to black ( Change color of
underlines as well) + iText
https://stackoverflow.com/questions/57993761/traverse-whole-pdf-and-change-blue-color-to-black-change-color-of-underlines-a
originalFile.pdf
https://raad-dev-test.s3.ap-south-1.amazonaws.com/36/2019-08-30/originalFile.pdf

testChangeFillRgbBlueToBlack - This canvas editor is related to the
editor in EditPageContent.testChangeBlackTextToGreenDocument(). In
contrast to that editor, though, the editor here shall change the fill
color of arbitrary content, not only text content. Thus, this actually
is simpler than the original.