New ChangeColor tests focusing on the stack overflow question
Traverse whole PDF and change blue color to black ( Change color of
underlines as well) + iText
https://stackoverflow.com/questions/57993761/traverse-whole-pdf-and-change-blue-color-to-black-change-color-of-underlines-a
021549Orig1s025_aprepitant_clinpharm_prea_Mac.pdf
https://raad-dev-test.s3.ap-south-1.amazonaws.com/36/2019-08-30/021549Orig1s025_aprepitant_clinpharm_prea_Mac.pdf
400_206494S5_avibactam_and_ceftazidine_unireview_prea_Mac.pdf
https://raad-dev-test.s3.ap-south-1.amazonaws.com/36/2019-08-30/400_206494S5_avibactam_and_ceftazidine_unireview_prea_Mac.pdf
Control_of_nitrosamine_impurities_in_sartans__rev.pdf
https://raad-dev-test.s3.ap-south-1.amazonaws.com/36/2019-08-30/Control_of_nitrosamine_impurities_in_sartans__rev.pdf
EDQM_reports_issues_of_non-compliance_with_tooth__Mac.pdf
https://raad-dev-test.s3.ap-south-1.amazonaws.com/36/2019-08-30/EDQM_reports_issues_of_non-compliance_with_tooth__Mac.pdf
originalFile.pdf
https://raad-dev-test.s3.ap-south-1.amazonaws.com/36/2019-08-30/originalFile.pdf

AllRgbBlueToBlackConverter - This PdfCanvasEditor replaces a large range
of blue'ish RGB colors by black. Essentially all colors are considered
blue in which the blue component is the strongest by a at least 10% and
is at least .5 itself. This matches a wide spectrum from purple to cyan
light enough to be recognized as such.
 All rg, RG, sc, SC, scn, and SCN instructions with three parameters all
of which have to be numeric are interpreted to set RGB colors. This
obviously is not correct in general but appears to match in case of the
OP's documents with DeviceRGB and ICCBased RGB color spaces.

testAllRgbBlueToBlackConverter* - These tests apply the
AllRgbBlueToBlackConverter to the test files.