New ImageDataRetrieval test focusing on the stack overflow question
PNG image is getting corrupted in Itext7
https://stackoverflow.com/questions/60990686/png-image-is-getting-corrupted-in-itext7
gist of image data URL
https://gist.github.com/nil96/d9d54200e5329211d2d7455492716500

testRetrieveLikeNil96 - Analyzing the data returned by
imageData.getData() shows that those data do not only contain -1 values
but others, too. But as the image is white with a bit of darker
scribble, there indeed is a lot of 255 (= -1), 254 (= -2), ... and only
a small amount of smaller unsigned values.