Test signing without reason and location values
Certain profiles prohibit the presence of the Reason entry in the signature dictionary.

testNullReasonAndLocation tries to create a signature without reason and location entries by using null values. Unfortunately it fails, iText signing always tries to set the reason and location.

testNullReasonAndLocationInEvent tries to create a signature without reason and location entries. It does so not by setting them to null in the signature appearance, though, but instead by registering a ISignatureEvent that removes the entries from the finished signature dictionary right before signing. This works.