New ObjectStructureAnalyzer utility plus test case
The ObjectStructureAnalyzer class tries to find all indirect objects in a PDF, also those in older revisions of a document replaced in the current revision, and even those referenced in no cross references.

It is in particular designed for detecting PDFs that might be used in hide-and-replace Shadow Attacks as proposed by the Ruhr Uni Bochum, see https://www.pdf-insecurity.org/signature-shadow/shadow-attacks.html

The AnalyzeStructure test uses the ObjectStructureAnalyzer to check whether there are multiple indirect objects with the same object number in the example PDFs from https://www.pdf-insecurity.org/signature-shadow/shadow-attacks.html provided as examples for Shadow Attacks. In the hide-and-replace example such objects are observed.