New SignWithAdaptions test focusing on the stack overflow question
iText Sign PDF with custom “Signed by” name
https://stackoverflow.com/questions/64678798/itext-sign-pdf-with-custom-signed-by-name

testSignWithNameInEvent - Adobe Acrobat uses the Name value in its signature panel if it cannot properly parse the signature container or if automatic validation during PDF opening is deactivated. Unfortunately not, though, if automatic validation is activated and the signature container can be parsed which is the most common case.