New MultipleSignaturesAndTagging test illustrating an issue with tagged PDFs and multiple signatures
testSignSimpleTaggedPdfOutlineTwiceCms - The first signature signs a PDF with an Outline element which points to a structure element which is changed by the second signature due to adding a reference to the second signature field to the structure tree. As a result Adobe Reader currently claims that the first signature is invalidated by disallowed changes.