New TextPosition test focusing on the stack overflow question
What value to use for .MoveUp of canvas
https://stackoverflow.com/questions/64972425/what-value-to-use-for-moveup-of-canvas

testForGustav - This test runs the code of the OP optimized to not rely on magic numbers: The offset by 4 observed by the OP actually is required to counteract the Paragraph default top margin value of 4. Thus, explicitly setting the paragraph top margin to 0 resolves the issue.