New OpenFile test focusing on the stack overflow question 66583475
PDF becoming corrupted after signing
https://stackoverflow.com/questions/66583475/pdf-becoming-corrupted-after-signing
xrefstreamdoc (17).pdf
https://ufile.io/uxq3gr6o

testOpenXrefstreamdoc17 - There are multiple issues in the cross reference stream of this file. In particular there are less entries (22) than claimed (23), there is a shift (entry for object 1 points to object 2 etc), and there are even more objects in the file (at least an object 24).