New ShowTextAtPosition test focusing on the stack overflow question 67601323
Add Paragraph via itext.Layout.Document.ShowTextAligned on first Page only
https://stackoverflow.com/questions/67601323/add-paragraph-via-itext-layout-document-showtextaligned-on-first-page-only

testAddCenteredBorderedParagraphOriginalFile - In the course of this question it became clear that the code in testAddCenteredBorderedParagraph() does not add to the first page but instead to the large. Using a different overload of Document.showTextAligned allows adding to the first page, though.