New CopyPdfsAndImages tests focusing on the stack overflow question 68377306
itext 7 (java) add image on a new page to the end of an existing pdf document
https://stackoverflow.com/questions/68377306/itext-7-java-add-image-on-a-new-page-to-the-end-of-an-existing-pdf-document

testMergeLikeAndreasHuber - This is a small condensed example of what can happen in the OP's original code. Indeed, images are added on pages of the imported PDFs.

testMergeLikeAndreasHuberImproved - This code now makes sure that images are drawn on a new page after the currently last one.