New AddFormFieldWeirdnesses test testSharedEmptyAnnotationArray
Given a two-page PDF in which the pages share an indirect object containing an empty array as annotations array. As the array is empty, this is not an invalid structure. Furthermore, it can realistically turn up as a result of an optimization.

Adding a form field to the second page of that PDF now results in the annotation of that field being present on both pages. This is neither valid nor desired.