New CreateSignatureAppearance test
testReuseSpecialAppearance - This test illustrates an issue in the ReuseAppearance feature of iText: Here the complete normal appearance of the unsigned field is re-used as n0 layer of the signed field. Unfortunately it was forgotten that the original appearance was displayed so that its BBox transformed by its matrix fits into the annotation rectangle. If BBox of the original appearance does not have its lower left corner in the origin or its matrix is not the identity, therefore, the re-used appearance usually is displayed in differently if at all.