New RotateText tests focusing on the stack overflow question 70786737
How to rotate a specific text in a pdf which is already present in the pdf using itext or pdfbox?
https://stackoverflow.com/questions/70786737/how-to-rotate-a-specific-text-in-a-pdf-which-is-already-present-in-the-pdf-using
Before.pdf as "BeforeAkhilNagaSai.pdf"
https://drive.google.com/file/d/1AUhsaWz3qwj5zbOgQ9r3cOAXDsHSLB1s/view?usp=sharing

testBeforeAkhilNagaSai - This test illustrates how to rotate text in an existing PDF using the LimitedTextRotater applied to the OP's example file.

testDocument - This test illustrates how to rotate text in an existing PDF using the LimitedTextRotater applied to another file. Here actually not merely a rotation but also a scaling is applied.