New SignBrokenPdf tests focusing on the stack overflow question 70847879
iText how to fix cross reference table for signed pdf
https://stackoverflow.com/questions/70847879/itext-how-to-fix-cross-reference-table-for-signed-pdf
testBrokenXref.pdf
broken by setting one entry to "0000000000 65536 n"

testSignBrokenXref - As expected iText throws an exception when trying to sign the given PDF with a broken cross reference entry in append mode.

testSignBrokenXrefForced - This test shows how to lead iText into believing the original file was not broken, so it can be signed in append mode.