New AddWatermark tests focusing on the stack overflow question 70889989
Pdf signature invalidates existing signature in Acrobat Reader
https://stackoverflow.com/questions/70889989/pdf-signature-invalidates-existing-signature-in-acrobat-reader
test.pdf as "testManuel.pdf"
https://easyupload.io/99fk3a

testSignTestManuelTwice - This reproduces the OP's issue in a single test method in Java. Just like the OP's .NET project and just like the test method testSignSimpleTaggedPdfOutlineTwiceCms(), the result with two signatures in Adobe Reader is negatively validated. The reason again is a change in the structure tree in combination with an outline entry pointing there.

testSignTestManuelTwiceNoTag - This reproduces the OP's issue in a single test method in Java except in one detail: structure tree manipulation during the second signature creation is prevented. And in contrast to testSignTestManuelTwice() (which reproduces the OP's .NET project behavior), the result with two signatures in Adobe Reader is positively validated.