New ExtractTextDimensions tests focusing on the stack overflow question 70789651
How to find the height and width of a text in a rotated pdf where the text is also rotated by same angle?
https://stackoverflow.com/questions/70789651/how-to-find-the-height-and-width-of-a-text-in-a-rotated-pdf-where-the-text-is-al
Before.pdf as "BeforeAkhilNagaSai.pdf"
https://drive.google.com/file/d/1AUhsaWz3qwj5zbOgQ9r3cOAXDsHSLB1s/view?usp=sharing

testBeforeAkhilNagaSai - Cannot reproduce the OP's issue. The width and height outputs are clearly not zero. Eventually it turned out that the problem exists only for older iText versions (7.1.14'ish).