New LargeCoordinateValues tests focusing on the Adobe community question 12739946
Number out of range error when gathering content from PDF content stream
https://community.adobe.com/t5/acrobat-sdk-discussions/number-out-of-range-error-when-gathering-content-from-pdf-content-stream/m-p/12739946

testCoordinateAt10000 - Using coordinate values in the 10000s nothing happens yet, but see testCoordinateAt100000.

testCoordinateAt100000 - Indeed, using coordinates in the 100000s, Adobe Reader rejects this, but see testCoordinateAt100000SmallBox.

testCoordinateAt100000SmallBox - In spite of using coordinates in the 100000s nothing happens. Apparently the issue is not with the coordinates in the content stream but with the media box coordinates (which here are small).