New CreateFormField .NET tests focusing on the stack overflow question 71166730
PdfTextFormField values is not getting populated with iText 7.2.0
https://stackoverflow.com/questions/71166730/pdftextformfield-values-is-not-getting-populated-with-itext-7-2-0

testCreateFieldLikeUser14463446 - This test contains the OP's code. As no on-page area is given, the field is created as invisible field.

testCreateFieldLikeUser14463446WithRectangle - This test contains the OP's code but using a different CreateText overload that also accepts a rectangle. Now the field is created as visble field, showing the field value.