New AddImage tests focusing on the stack overflow question
iText - add image to an existing PDF
https://stackoverflow.com/questions/71488649/itext-add-image-to-an-existing-pdf
step1.pdf
https://ipupload.com/tP6/step1.pdf

testLikeKalpit - Indeed, this creates an invalid PDF as the changes to the resources, the added image resource, are not stored in the result file. The cause is that the addition to the resources only marks the resources and the specific resource type dictionary as changed (both of which are direct objects in the case at hand) and the contents array already is indirect, so changes to it don't mark the page changed, either.

testLikeKalpitPlusPage - In contrast to testLikeKalpit we here explicitly mark the page object in question as changed. As a consequence, the changes in the resources also are stored and all is ok.