New SignDeferred test focusing on the stack overflow question 72578658
Invalidating the signature when the pdf contains image, but works fine without the image in itext
https://stackoverflow.com/questions/72578658/invalidating-the-signature-when-the-pdf-contains-image-but-works-fine-without-t

testSignLikeNikhil - This test should have tested the OP's signing issue. Unfortunately their code is full of references to unknown variables and types. Thus, currently this test only generates a PDF using the OP's code for generating a document. Signing this results in a properly signing file. Thus, the issue cannot be reproduced.